Add extra variants of ghost cell routines
This makes it possible to leave out the refinement boundary and physical
boundary method if they are already stored and to fill ghost cells for multiple
variables at the same time.

A variable tree%cc_methods_vars has also been introduced, to keep track of the
variables for which methods have been stored.